+++ crates/api-desc/CHANGELOG.md
@@ -13,3 +13,5 @@
 - Add copyright header on generated files
 
 ## 0.1.0
+
+<!-- Update PR number to skip CHANGELOG.md test: #0 -->