+++ crates/api-desc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 0.1.2-git
+
 ### Minor
 
 - Add scheduling::breakpoint